cleanup and consolidation
- made acPlugins4net.dll signed
- set version of acPlugins4net.dll to 2.0: the major version should match the acServer UDP protocol version, minor should be changed at least for each release
- AcServerPluginManager: removed ServerName, Track, TrackLayout Properties (use MsgSessionInfo instead)
- AcServerPlugin: removed some obsolete properties
- app.config added load_server_cfg (default 1): set to 0 when plugin is run as an external plugin
   in this case ac_server_port, ac_server_host, plugin_port and max_clients should be set manually in app.config
- app.config added max_clients, only needed if settings are not loaded from server cfg (plugin run as external plugin)
- adjusted app.config of Admin2ChatPlugin and MinoRatingPlugin to include all settings that can be set
- adjusted Main methods of Admin2ChatPlugin and MinoRatingPlugin, so that additional plugin configured in app.config are loaded
- changed Admin2ChatPlugin to .NET 4.0 Client Profile

+++ acplugins4net/acServerFake/viemodel/messages/NewSessionViewModel.cs
@@ -85,7 +85,7 @@ namespace acServerFake.viemodel.messages
         public NewSessionViewModel()
         {
             // Quick & dirty: Some defaults to speed up tests
-            Message.Version = AcServerPluginManager.RequiredProtocolVersion - 1;
+            Message.Version = (byte)(AcServerPluginManager.RequiredProtocolVersion - 1);
             ServerName = "Fake Server Trackday";
             Track = "mugello";
             TrackConfig = "";
